update delete method in previousworkHistoryServiseImpl

+++ src/main/java/lk/zerocode/api/service/impl/PreviousWorkHistoryServiceImpl.java
@@ -2,11 +2,13 @@ package lk.zerocode.api.service.impl;
 
 import lk.zerocode.api.controller.request.EmergencyContactRequest;
 import lk.zerocode.api.controller.request.PreviousWorkHistoryRequest;
+import lk.zerocode.api.controller.response.DependentDetailMsgResponse;
 import lk.zerocode.api.controller.response.EmergencyResponse;
 import lk.zerocode.api.controller.response.PreviousWorkHistoryIdResponse;
 import lk.zerocode.api.controller.response.PreviousWorkHistoryResponse;
 import lk.zerocode.api.exceptions.EmployeeNotFoundException;
 import lk.zerocode.api.exceptions.PreviousWorkHistoryNotFoundException;
+import lk.zerocode.api.model.DependentDetail;
 import lk.zerocode.api.model.EmergencyContact;
 import lk.zerocode.api.model.Employee;
 import lk.zerocode.api.model.PreviousWorkHistory;
@@ -65,9 +67,6 @@ public class PreviousWorkHistoryServiceImpl implements PreviousWorkHistoryServic
 
 
 
-
-
-
     @Override
     public List<PreviousWorkHistoryResponse> getPreviousWorkHistoryByEmpId(Long eid) throws EmployeeNotFoundException {
         Optional<Employee> employeeOptional = employeeRepository.findById(eid);
@@ -113,7 +112,7 @@ public class PreviousWorkHistoryServiceImpl implements PreviousWorkHistoryServic
         Optional<PreviousWorkHistory> previousWorkHistoryOptional = previousWorkHistoryRepository.findById(previousWorkHistoryId);
 
         if (!previousWorkHistoryOptional.isPresent()) {
-            throw new PreviousWorkHistoryNotFoundException("Previous Work History not found with ID: " + previousWorkHistoryId);
+            throw new PreviousWorkHistoryNotFoundException("Previous Work History not found with ID : " + previousWorkHistoryId);
         }
 
         PreviousWorkHistory previousWorkHistory = previousWorkHistoryOptional.get();
@@ -138,21 +137,24 @@ public class PreviousWorkHistoryServiceImpl implements PreviousWorkHistoryServic
 
 
     @Override
-    public String deletePreviousWorkHistoryDetailsById(Long empId, Long previousWorkHistoryId) throws EmployeeNotFoundException{
+    public String deletePreviousWorkHistoryDetailsById(Long empId, Long previousWorkHistoryId) throws EmployeeNotFoundException,PreviousWorkHistoryNotFoundException{
 
         Optional<Employee> employeeOptional = employeeRepository.findById(empId);
         if (!employeeOptional.isPresent()) {
             throw new EmployeeNotFoundException("Employee not found with id: " + empId);
         }
 
-        Optional<PreviousWorkHistory> contactOptional = previousWorkHistoryRepository.findById(previousWorkHistoryId);
-//        if (!contactOptional.isPresent()) {
-//            return null;
-//        }
+        Employee employee =employeeOptional.get();
 
+        Optional<PreviousWorkHistory> previousWorkHistoryOptional =previousWorkHistoryRepository.findPreviousWorkHistoriesByEmployeeAndId(employee,previousWorkHistoryId);
+        if(!previousWorkHistoryOptional.isPresent()){
+            throw new PreviousWorkHistoryNotFoundException("Employee not found with id: "+previousWorkHistoryId);
+        }
+        PreviousWorkHistory previousWorkHistory =previousWorkHistoryOptional.get();
         previousWorkHistoryRepository.deleteById(previousWorkHistoryId);
-        return "Delete successful for employee id: " + empId + " and contact id: " + previousWorkHistoryId;
-    }
 
+        return ("Delete successfully previousWorkHistoryId no "+previousWorkHistory.getId()+" of employee with Id: "+empId);
+
+    }
 
 }
